Move event templates into events folder
