+++ src/pages/events/event-templates/coding-for-everyone/coding-for-everyone.json
