+++ src/pages/events/event-templates/event-templates.json
