+++ src/pages/events/event-templates/short-course/short-course.json
